+++ packages/gov-ui-kit/src/modules/components/asset/assetDataListItem/assetDataListItemStructure/assetDataListItemStructure.test.tsx
@@ -20,53 +20,31 @@ describe('<AssetDataListItem.Structure /> component', () => {
         );
     };
 
-    it('renders tokenName and symbol', () => {
-        const props = {
-            name: 'Ethereum',
-            symbol: 'ETH',
-            amount: 420.69,
-        };
-
+    it('renders token name and symbol', () => {
+        const props = { name: 'Ethereum', symbol: 'ETH' };
         render(createTestComponent(props));
         expect(screen.getByText(props.name)).toBeInTheDocument();
         expect(screen.getByText(props.symbol)).toBeInTheDocument();
     });
 
-    it('renders amount, fiat price', async () => {
-        const props = {
-            name: 'Ethereum',
-            symbol: 'ETH',
-            amount: 420.69,
-            fiatPrice: 3654.76,
-        };
-
+    it('correctly renders amount, fiat price and price change', () => {
+        const props = { amount: 10, fiatPrice: 1250, priceChange: 25 };
         render(createTestComponent(props));
-        const USDAmount = await screen.findByText(/1.54/);
-        expect(USDAmount).toHaveTextContent('$1.54M');
+        expect(screen.getByText('$12.50K')).toBeInTheDocument();
         expect(screen.getByText(props.amount)).toBeInTheDocument();
+        expect(screen.getByText('+$2.50K')).toBeInTheDocument();
+        expect(screen.getByText('+25.00%')).toBeInTheDocument();
     });
 
-    it('handles not passing fiat price', () => {
-        const props = {
-            name: 'Ethereum',
-            symbol: 'ETH',
-            amount: 0,
-            priceChange: 0,
-        };
-
+    it('renders unknown with fiatPrice is not set', () => {
+        const props = { fiatPrice: undefined };
         render(createTestComponent(props));
-        expect(screen.getByText('Unknown')).toBeInTheDocument(); // Assuming Tag component renders '0%' for zero priceChange
+        expect(screen.getByText('Unknown')).toBeInTheDocument();
     });
 
-    it('handle not passing priceChange', async () => {
-        const props = {
-            name: 'Ethereum',
-            amount: 420.69,
-            symbol: 'ETH',
-            fiatPrice: 3654.76,
-        };
-
+    it('correctly renders price change when fiatPrice is set but priceChange property is undefined', () => {
+        const props = { priceChange: undefined, fiatPrice: 10 };
         render(createTestComponent(props));
-        expect(screen.getByText('0%')).toBeInTheDocument();
+        expect(screen.getByText('0.00%')).toBeInTheDocument();
     });
 });