+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.45] - 2024-08-23
+
 ### Added
 
 -   Update illustration SVGs with custom CSS color customizations